requirements.txt
fastapi
uvicorn[standard]
duckduckgo_search
requests
beautifulsoup4
transformers
torch
python-dotenv
